Finall(?)

+++ BA/QLBaiDoXe/QLBaiDoXe/DSThe.xaml.cs
@@ -1,6 +1,7 @@
 ﻿using System.Windows;
 using System.Windows.Controls;
 using QLBaiDoXe.DBClasses;
+using QLBaiDoXe.ParkingLotModel;
 
 namespace QLBaiDoXe
 {
@@ -35,11 +36,18 @@ namespace QLBaiDoXe
             }
             else
             {
+                
+
                 if (MessageBox.Show("Bạn có muốn xóa thẻ đã chọn?", "Xác nhận", MessageBoxButton.YesNo) == MessageBoxResult.No)
                     return;
                 var selectedItems = (dynamic)ListThe.SelectedItems[0];
+                if ( Cards.CheckCardState((long)selectedItems.ParkingCardID) == 1)
+                {
+                    MessageBox.Show("Thẻ đang được sử dụng", "Lỗi!");
+                    return;
+                }
                 Cards.DeleteCard((long)selectedItems.ParkingCardID);
-                MessageBox.Show("Đã xóa thẻ thành công!");
+                MessageBox.Show("Đã xóa thẻ thành công!", "Thông báo!");
                 ListThe.ItemsSource = null;
                 ListThe.ItemsSource = Cards.GetAllParkingCards();
             }
